chore: reorganize unit tests in the local package
