+++ packages/local/test/unit/commands/local/index.unit.test.ts
@@ -1,6 +1,6 @@
 import {expect, test} from '@oclif/test'
 
-import * as foreman from '../../../src/fork-foreman'
+import * as foreman from '../../../../src/fork-foreman'
 
 // eslint-disable-next-line node/no-missing-require
 const procfile = require('../../../src/load-foreman-procfile')
